@@ -365,8 +365,14 @@ func (pmm *PMM) Restart(ctx context.Context) error {
 
 	tCtx, cancel := context.WithTimeout(ctx, getTimeout)
 	defer cancel()
-	if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/server/version"); err != nil && !errors.Is(err, io.EOF) {
-		return errors.Wrap(err, "failed to ping PMM")
+	if pmm.pmmVersion[0:1] == "2" {
+		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/version"); err != nil && !errors.Is(err, io.EOF) {
+			return errors.Wrap(err, "failed to ping PMM")
+		}
+	} else if pmm.pmmVersion[0:1] == "3" {
+		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/server/version"); err != nil && !errors.Is(err, io.EOF) {
+			return errors.Wrap(err, "failed to ping PMM")
+		}
 	}
 	return nil
 }
