@@ -48,10 +48,18 @@ func TestExportImport(t *testing.T) {
 	var b util.Binary
 	testDir := t.TempDir()
 
-	args := []string{"-d", filepath.Join(testDir, "dump.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-qan", "--click-house-url", pmm.ClickhouseURL()}
+	pmm.Log("Checking filtering with `--instance` flag")
+	args := []string{"-d", filepath.Join(testDir, "filter-dump.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-qan", "--click-house-url", pmm.ClickhouseURL(), "--instance", "pmm-server"}
+	stdout, stderr, err := b.Run(append([]string{"export", "--ignore-load"}, args...)...)
+	if err != nil {
+		t.Fatal("failed to export", err, stdout, stderr)
+	}
+	checkDumpFiltering(t, filepath.Join(testDir, "filter-dump.tar.gz"), "pmm-server")
+
+	args = []string{"-d", filepath.Join(testDir, "dump.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-qan", "--click-house-url", pmm.ClickhouseURL()}
 
 	pmm.Log("Exporting data to", filepath.Join(testDir, "dump.tar.gz"))
-	stdout, stderr, err := b.Run(append([]string{"export", "--ignore-load"}, args...)...)
+	stdout, stderr, err = b.Run(append([]string{"export", "--ignore-load"}, args...)...)
 	if err != nil {
 		t.Fatal("failed to export", err, stdout, stderr)
 	}
